Clean up goals for web repos and others
Remove use of allSatisfied, it is redundant since whenPushSatisfies
ensures all of its guards are satisfied.

Move to pre-approval for web site deployments.

Only deploy sites when commit is on default branch.

Remove release changelog goal from web-app.

Simplify pathToIndex so the result link points to the root of the
site.

Update sdm-pack-issue.